Design: fix CommunityFragment design
delete CommunityAllFragment and add design to Report/Talk Fragment
