+++ app/src/main/java/com/umc/approval/ui/fragment/community/CommunityReportFragment.kt
@@ -5,12 +5,20 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.umc.approval.data.dto.community.CommunityItemDto
+import com.umc.approval.data.dto.opengraph.OpenGraphDto
 import com.umc.approval.databinding.FragmentCommunityReportBinding
+import com.umc.approval.ui.adapter.community_fragment.CommunityReportItemRVAdapter
 
 class CommunityReportFragment : Fragment() {
 
     private var _binding : FragmentCommunityReportBinding? = null
     private val binding get() = _binding!!
+
+    //Community Image RV Adapter
+    private lateinit var communityReportItemRVAdapter: CommunityReportItemRVAdapter
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -24,7 +32,84 @@ class CommunityReportFragment : Fragment() {
         _binding = FragmentCommunityReportBinding.inflate(inflater, container, false)
         val view = binding.root
 
+        connect_to_community_rv()
+
         return view
+    }
+
+    private fun connect_to_community_rv() {
+        val init_data = mutableListOf<CommunityItemDto>()
+
+        var openGraphDto = OpenGraphDto(
+            "https://www.naver.com/",
+            "네이버",
+            "네이버",
+            "네이버",
+            "https://s.pstatic.net/static/www/mobile/edit/2016/0705/mobile_212852414260.png"
+        )
+        init_data.add(
+            CommunityItemDto
+                (
+                0, "", "", "", "", "", "",
+                "", "", "", "", mutableListOf(), openGraphDto, 0, 0, 0, 0
+            )
+        )
+        init_data.add(
+            CommunityItemDto
+                (
+                1, "", "", "", "", "", "",
+                "", "", "", "", mutableListOf(), openGraphDto, 0, 0, 0, 0
+            )
+        )
+        init_data.add(
+            CommunityItemDto
+                (
+                0, "", "", "", "", "", "",
+                "", "", "", "", mutableListOf(), openGraphDto, 0, 0, 0, 0
+            )
+        )
+        init_data.add(
+            CommunityItemDto
+                (
+                1, "", "", "", "", "", "",
+                "", "", "", "", mutableListOf(), openGraphDto, 0, 0, 0, 0
+            )
+        )
+        init_data.add(
+            CommunityItemDto
+                (
+                0, "", "", "", "", "", "",
+                "", "", "", "", mutableListOf(), openGraphDto, 0, 0, 0, 0
+            )
+        )
+        init_data.add(
+            CommunityItemDto
+                (
+                1, "", "", "", "", "", "",
+                "", "", "", "", mutableListOf(), openGraphDto, 0, 0, 0, 0
+            )
+        )
+        init_data.add(
+            CommunityItemDto
+                (
+                0, "", "", "", "", "", "",
+                "", "", "", "", mutableListOf(), openGraphDto, 0, 0, 0, 0
+            )
+        )
+        init_data.add(
+            CommunityItemDto
+                (
+                0, "", "", "", "", "", "",
+                "", "", "", "", mutableListOf(), openGraphDto, 0, 0, 0, 0
+            )
+        )
+
+        communityReportItemRVAdapter = CommunityReportItemRVAdapter(init_data)
+
+        val community_item_rv: RecyclerView = binding.communityRvItem
+
+        community_item_rv.adapter = communityReportItemRVAdapter
+        community_item_rv.layoutManager = LinearLayoutManager(requireContext(), LinearLayoutManager.VERTICAL, false)
     }
 
     /**
