+++ app/src/main/res/layout/fragment_community_report.xml
@@ -1,14 +1,88 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".ui.fragment.community.CommunityReportFragment">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
+    <com.google.android.material.chip.ChipGroup
+        android:id="@+id/community_category"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:layout_height="wrap_content"
+        android:layout_marginTop="10dp"
+        android:layout_marginLeft="10dp"
+        app:layout_constraintTop_toTopOf="parent"
+        app:selectionRequired="true"
+        app:singleLine="true"
+        app:singleSelection="true">
 
-</FrameLayout>
+        <com.google.android.material.chip.Chip
+            android:id="@+id/new_category"
+            style="@style/Colors_Widget.MaterialComponents.Chip.Choice"
+            android:layout_width="wrap_content"
+            android:layout_height="22dp"
+            android:checked="true"
+            android:gravity="center"
+            android:text="최신"
+            android:textSize="11dp"
+            app:chipCornerRadius="15dp"
+            app:chipEndPadding="2dp"
+            app:chipMinTouchTargetSize="0dp"
+            app:chipStartPadding="0dp" />
+
+        <com.google.android.material.chip.Chip
+            android:id="@+id/hot_category"
+            style="@style/Colors_Widget.MaterialComponents.Chip.Choice"
+            android:layout_width="wrap_content"
+            android:layout_height="22dp"
+            android:gravity="center"
+            android:text="인기"
+            android:textSize="11dp"
+            app:chipCornerRadius="15dp"
+            app:chipEndPadding="2dp"
+            app:chipMinTouchTargetSize="0dp"
+            app:chipStartPadding="0dp" />
+
+        <com.google.android.material.chip.Chip
+            android:id="@+id/follow_category"
+            style="@style/Colors_Widget.MaterialComponents.Chip.Choice"
+            android:layout_width="wrap_content"
+            android:layout_height="22dp"
+            android:gravity="center"
+            android:text="디지털기기"
+            android:textSize="11dp"
+            app:chipCornerRadius="15dp"
+            app:chipEndPadding="2dp"
+            app:chipMinTouchTargetSize="0dp"
+            app:chipStartPadding="0dp" />
+
+        <com.google.android.material.chip.Chip
+            android:id="@+id/my_category"
+            style="@style/Colors_Widget.MaterialComponents.Chip.Choice"
+            android:layout_width="wrap_content"
+            android:layout_height="22dp"
+            android:gravity="center"
+            android:text="디지털기기"
+            android:textSize="11sp"
+            app:chipCornerRadius="15dp"
+            app:chipEndPadding="2dp"
+            app:chipMinTouchTargetSize="0dp"
+            app:chipStartPadding="0dp" />
+    </com.google.android.material.chip.ChipGroup>
+
+    <ScrollView
+        android:layout_marginTop="15dp"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        app:layout_constraintTop_toBottomOf="@+id/community_category">
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/community_rv_item"
+            android:layout_marginBottom="10dp"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+        </androidx.recyclerview.widget.RecyclerView>
+    </ScrollView>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
